feat: Add optional name-only flag to git diff command
This commit adds support for an optional name-only flag to the git diff
command. If the flag is provided as an argument, the command will only
output the names of the modified files, instead of the full diff. This
can be useful for quickly identifying modified files without viewing the
full changes